Add files via upload
DON"T RUN THIS. It's attached to my billing account and will cost moneyyyy. I don't even think it can run without my JSON key (and actual query is commented out just to be safe) but still... 

Anyways this is my WIP for the sentiment analysis for TSLA using the GDELT event data set. Query'd the data with a python SQL query through google's Big Query and converted it to a Pandas data frame.  Then I saved as a parquet and csv. 

CSV is also attached here.